Deleted the test file created under tests/test_cmdline, moved the test cases to tests/test_commands.py
